+++ apps/astarte_realm_management/config/dev.exs
@@ -2,3 +2,15 @@ use Mix.Config
 
 config :cqerl,
   cassandra_nodes: [{System.get_env("CASSANDRA_DB_HOST"), System.get_env("CASSANDRA_DB_PORT")}]
+
+config :logger, :console,
+  format: {PrettyLog.LogfmtFormatter, :format},
+  metadata: [
+    :realm,
+    :interface,
+    :interface_major,
+    :trigger_name,
+    :module,
+    :function,
+    :tag
+  ]