+++ apps/astarte_realm_management/config/test.exs
@@ -7,3 +7,11 @@ config :cqerl,
   ]
 
 config :astarte_rpc, :amqp_connection, host: System.get_env("RABBITMQ_HOST") || "rabbitmq"
+
+config :logger, :console,
+  format: {PrettyLog.UserFriendlyFormatter, :format},
+  metadata: [
+    :realm,
+    :function,
+    :tag
+  ]