+++ apps/astarte_realm_management/lib/astarte_realm_management/engine.ex
@@ -37,6 +37,8 @@ defmodule Astarte.RealmManagement.Engine do
   alias CQEx.Client, as: DatabaseClient
 
   def get_health() do
+    _ = Logger.debug("Get health.")
+
     with {:ok, client} <- Database.connect(),
          :ok <- Queries.check_astarte_health(client, :each_quorum) do
       {:ok, %{status: :ready}}
@@ -59,7 +61,7 @@ defmodule Astarte.RealmManagement.Engine do
   end
 
   def install_interface(realm_name, interface_json, opts \\ []) do
-    Logger.debug("Going to install a new interface on realm #{realm_name}.")
+    _ = Logger.info("Going to install a new interface.", tag: "install_interface")
 
     with {:ok, client} <- Database.connect(realm_name),
          {:ok, json_obj} <- Jason.decode(interface_json),
@@ -71,9 +73,15 @@ defmodule Astarte.RealmManagement.Engine do
            {:interface_avail, Queries.is_interface_major_available?(client, name, major)},
          :ok <- Queries.check_correct_casing(client, name),
          {:ok, automaton} <- EndpointsAutomaton.build(interface_doc.mappings) do
-      Logger.info("Installing interface #{name} v#{Integer.to_string(major)} on #{realm_name}.")
+      _ =
+        Logger.info("Installing interface.",
+          interface: name,
+          interface_major: major,
+          tag: "install_interface_started"
+        )
 
       if opts[:async] do
+        # TODO: add _ = Logger.metadata(realm: realm_name)
         Task.start(Queries, :install_new_interface, [client, interface_doc, automaton])
 
         {:ok, :started}
@@ -82,11 +90,19 @@ defmodule Astarte.RealmManagement.Engine do
       end
     else
       {:error, {:invalid, _invalid_str, _invalid_pos}} ->
-        Logger.warn("Received invalid interface JSON: #{inspect(interface_json)}")
+        _ =
+          Logger.warn("Received invalid interface JSON: #{inspect(interface_json)}.",
+            tag: "invalid_json"
+          )
+
         {:error, :invalid_interface_document}
 
       {:error, %Ecto.Changeset{} = changeset} ->
-        Logger.warn("Received invalid interface: #{inspect(changeset)}")
+        _ =
+          Logger.warn("Received invalid interface: #{inspect(changeset)}.",
+            tag: "invalid_interface_document"
+          )
+
         {:error, :invalid_interface_document}
 
       {:error, :database_connection_error} ->
@@ -107,7 +123,7 @@ defmodule Astarte.RealmManagement.Engine do
   end
 
   def update_interface(realm_name, interface_json, opts \\ []) do
-    Logger.debug("Going to update an interface on realm #{realm_name}.")
+    _ = Logger.info("Going to perform interface update.", tag: "update_interface")
 
     with {:ok, client} <- Database.connect(realm_name),
          {:ok, json_obj} <- Jason.decode(interface_json),
@@ -123,8 +139,6 @@ defmodule Astarte.RealmManagement.Engine do
          :ok <- error_on_downgrade(installed_interface, interface_descriptor),
          {:ok, new_mappings} <- extract_new_mappings(client, interface_doc),
          {:ok, automaton} <- EndpointsAutomaton.build(interface_doc.mappings) do
-      Logger.info("Updating interface #{name} v#{Integer.to_string(major)} on #{realm_name}.")
-
       new_mappings_list = Map.values(new_mappings)
 
       interface_update =
@@ -133,6 +147,7 @@ defmodule Astarte.RealmManagement.Engine do
         end)
 
       if opts[:async] do
+        # TODO: add _ = Logger.metadata(realm: realm_name)
         Task.start_link(__MODULE__, :execute_interface_update, [
           client,
           interface_update,
@@ -155,11 +170,19 @@ defmodule Astarte.RealmManagement.Engine do
       end
     else
       {:error, {:invalid, _invalid_str, _invalid_pos}} ->
-        Logger.warn("Received invalid interface JSON: #{inspect(interface_json)}")
+        _ =
+          Logger.warn("Received invalid interface JSON: #{inspect(interface_json)}.",
+            tag: "invalid_json"
+          )
+
         {:error, :invalid_interface_document}
 
       {:error, %Ecto.Changeset{} = changeset} ->
-        Logger.warn("Received invalid interface: #{inspect(changeset)}")
+        _ =
+          Logger.warn("Received invalid interface: #{inspect(changeset)}.",
+            tag: "invalid_interface_document"
+          )
+
         {:error, :invalid_interface_document}
 
       {:error, :database_connection_error} ->
@@ -192,6 +215,16 @@ defmodule Astarte.RealmManagement.Engine do
   end
 
   def execute_interface_update(client, interface_descriptor, new_mappings, automaton, descr, doc) do
+    name = interface_descriptor.name
+    major = interface_descriptor.major_version
+
+    _ =
+      Logger.info("Updating interface.",
+        interface: name,
+        interface_major: major,
+        tag: "update_interface_started"
+      )
+
     with :ok <- Queries.update_interface_storage(client, interface_descriptor, new_mappings) do
       Queries.update_interface(client, interface_descriptor, new_mappings, automaton, descr, doc)
     end
@@ -234,7 +267,7 @@ defmodule Astarte.RealmManagement.Engine do
       :ok
     else
       incompatible_value ->
-        Logger.debug("Incompatible change: #{inspect(incompatible_value)}")
+        _ = Logger.debug("Incompatible change: #{inspect(incompatible_value)}.")
         {:error, :invalid_update}
     end
   end
@@ -280,7 +313,12 @@ defmodule Astarte.RealmManagement.Engine do
   end
 
   def delete_interface(realm_name, name, major, opts \\ []) do
-    Logger.debug("Going to delete #{name} v#{Integer.to_string(major)} on #{realm_name}.")
+    _ =
+      Logger.info("Going to delete interface.",
+        tag: "delete_interface",
+        interface: name,
+        interface_major: major
+      )
 
     with {:major, 0} <- {:major, major},
          {:ok, client} <- Database.connect(realm_name),
@@ -292,6 +330,7 @@ defmodule Astarte.RealmManagement.Engine do
          {:triggers, {:ok, false}} <-
            {:triggers, Queries.has_interface_simple_triggers?(client, interface_id)} do
       if opts[:async] do
+        # TODO: add _ = Logger.metadata(realm: realm_name)
         Task.start_link(Engine, :execute_interface_deletion, [client, name, major])
 
         {:ok, :started}
@@ -324,11 +363,24 @@ defmodule Astarte.RealmManagement.Engine do
          {:ok, descriptor} <- InterfaceDescriptor.from_db_result(interface_row),
          :ok <- Queries.delete_interface_storage(client, descriptor),
          :ok <- Queries.delete_devices_with_data_on_interface(client, name) do
+      _ =
+        Logger.info("Interface deletion started.",
+          interface: name,
+          interface_major: major,
+          tag: "delete_interface_started"
+        )
+
       Queries.delete_interface(client, name, major)
     end
   end
 
   def interface_source(realm_name, interface_name, major_version) do
+    _ =
+      Logger.debug("Get interface source.",
+        interface: interface_name,
+        interface_major: major_version
+      )
+
     with {:ok, client} <- Database.connect(realm_name),
          {:ok, interface} <- Queries.fetch_interface(client, interface_name, major_version) do
       Jason.encode(interface)
@@ -336,6 +388,8 @@ defmodule Astarte.RealmManagement.Engine do
   end
 
   def list_interface_versions(realm_name, interface_name) do
+    _ = Logger.debug("List interface versions.", interface: interface_name)
+
     with {:ok, client} <- Database.connect(realm_name) do
       Queries.interface_available_versions(client, interface_name)
     else
@@ -348,6 +402,8 @@ defmodule Astarte.RealmManagement.Engine do
   end
 
   def get_interfaces_list(realm_name) do
+    _ = Logger.debug("Get interfaces list.")
+
     with {:ok, client} <- Database.connect(realm_name) do
       Queries.get_interfaces_list(client)
     else
@@ -360,6 +416,8 @@ defmodule Astarte.RealmManagement.Engine do
   end
 
   def get_jwt_public_key_pem(realm_name) do
+    _ = Logger.debug("Get JWT public key PEM.")
+
     with {:ok, client} <-
            DatabaseClient.new(
              List.first(Application.get_env(:cqerl, :cassandra_nodes)),
@@ -378,6 +436,7 @@ defmodule Astarte.RealmManagement.Engine do
              List.first(Application.get_env(:cqerl, :cassandra_nodes)),
              keyspace: realm_name
            ) do
+      _ = Logger.info("Updating JWT public key PEM.", tag: "updating_jwt_pub_key")
       Queries.update_jwt_public_key_pem(client, jwt_public_key_pem)
     else
       {:error, :shutdown} ->
@@ -386,6 +445,12 @@ defmodule Astarte.RealmManagement.Engine do
   end
 
   def install_trigger(realm_name, trigger_name, action, serialized_tagged_simple_triggers) do
+    _ =
+      Logger.info("Going to install a new trigger.",
+        trigger_name: trigger_name,
+        tag: "install_trigger"
+      )
+
     with {:ok, client} <- get_database_client(realm_name),
          {:exists?, {:error, :trigger_not_found}} <-
            {:exists?, Queries.retrieve_trigger_uuid(client, trigger_name)},
@@ -396,6 +461,12 @@ defmodule Astarte.RealmManagement.Engine do
          :ok <- validate_simple_triggers(client, simple_trigger_maps),
          # TODO: these should be batched together
          :ok <- install_simple_triggers(client, simple_trigger_maps, trigger_uuid, target) do
+      _ =
+        Logger.info("Installing trigger.",
+          trigger_name: trigger_name,
+          tag: "install_trigger_started"
+        )
+
       Queries.install_trigger(client, trigger)
     else
       {:exists?, _} ->
@@ -528,6 +599,8 @@ defmodule Astarte.RealmManagement.Engine do
   end
 
   def get_trigger(realm_name, trigger_name) do
+    _ = Logger.debug("Get trigger.", trigger_name: trigger_name)
+
     with {:ok, client} <- get_database_client(realm_name),
          {:ok, %Trigger{} = trigger} <- Queries.retrieve_trigger(client, trigger_name) do
       %Trigger{
@@ -567,14 +640,21 @@ defmodule Astarte.RealmManagement.Engine do
   end
 
   def get_triggers_list(realm_name) do
+    _ = Logger.debug("Get triggers list.")
+
     with {:ok, client} <- get_database_client(realm_name) do
       Queries.get_triggers_list(client)
     end
   end
 
   def delete_trigger(realm_name, trigger_name) do
+    _ = Logger.info("Going to delete trigger.", trigger_name: trigger_name, tag: "delete_trigger")
+
     with {:ok, client} <- get_database_client(realm_name),
          {:ok, trigger} <- Queries.retrieve_trigger(client, trigger_name) do
+      _ =
+        Logger.info("Deleting trigger.", trigger_name: trigger_name, tag: "delete_trigger_started")
+
       delete_all_succeeded =
         Enum.all?(trigger.simple_triggers_uuids, fn simple_trigger_uuid ->
           Queries.delete_simple_trigger(client, trigger.trigger_uuid, simple_trigger_uuid) == :ok